Tests/LibWeb/Text/expected/async-scrolling/nested-navigable-wheel-admission.txt

main thread regions: 1
over iframe: blocked-by-main-thread-region
outside iframe: accepted
